changes after QA
