+++ src/components/MyAccountConsents/MyAccountConsents.js
@@ -4,6 +4,7 @@ import PropTypes from 'prop-types';
 import { withTranslation } from 'react-i18next';
 import Loader from 'components/Loader';
 import submitConsents from 'api/Customer/submitConsents';
+import translateConsents from 'util/consentsHelper';
 import {
   ButtonStyled,
   CheckboxStyled,
@@ -117,7 +118,7 @@ class MyAccountConsents extends Component {
             required={item.required}
             hide={showConsentsOnly && !item.required}
           >
-            {t(item.label)}
+            {translateConsents(item.label, t)}
           </CheckboxStyled>
         ))}
         {!showConsentsOnly && (
