+++ src/translations/en/translations.json
@@ -41,6 +41,7 @@
   "Please agree on all required consents to use this service": "Please agree on all required consents to use this service",
   "I accept the {{htmltag}}Terms and Conditions{{endhtmltag}} of Cleeng.": "I accept the {{htmltag}}Terms and Conditions{{endhtmltag}} of Cleeng.",
   "I accept {{htmltag}}Terms and Conditions{{endhtmltag}} of Cleeng.": "I accept {{htmltag}}Terms and Conditions{{endhtmltag}} of Cleeng.",
+  "Yes, I want to receive Cleeng updates by email.": "Yes, I want to receive Cleeng updates by email.",
   "Go to register": "Go to register",
   "Google": "Google",
   "Facebook": "Facebook",
@@ -127,8 +128,11 @@
   "This term is mandatory for using myAccount": "This term is mandatory for using myAccount",
   "Edit Password": "Edit Password",
   "If you want to edit your password, click 'YES, Reset' to receive password reset instruction on your mail": "If you want to edit your password, click 'YES, Reset' to receive password reset instruction on your mail",
+  "Email has been sent!": "Email has been sent!",
+  "Please check your inbox and check the Instruction to change a password": "Please check your inbox and check the Instruction to change a password",
   "Terms and Conditions": "Terms and Conditions",
   "Terms & Conditions": "Terms & Conditions",
+  "Please accept Terms & Conditions": "Please accept Terms & Conditions",
   "Hi there!": "Hi there!",
   "Your brand new account is waiting for you. Just one last step. We would like you to review our terms and conditions before continuing to your account.": "Your brand new account is waiting for you. Just one last step. We would like you to review our terms and conditions before continuing to your account.",
   "Let’s do it": "Let’s do it",
@@ -195,8 +199,6 @@
   "Your profile details have been changed successfully": "Your profile details have been changed successfully",
   "Your address details have been changed successfully": "Your address details have been changed successfully",
   "Something went wrong. Try again later.": "Something went wrong. Try again later.",
-  "Address Line 1": "Address Line 1",
-  "Address Line 2": "Address Line 2",
   "State": "State",
   "Zip/Postal Code": "Zip/Postal Code",
   "Edit Address": "Edit Address",
@@ -223,6 +225,7 @@
   "Add your card details here": "Add your card details here",
   "Connect your PayPal account here": "Connect your PayPal account here",
   "Remove your payment method": "Remove your payment method",
+  "Your new details will replace the details used for your other active subscriptions.": "Nowe dane zastąpią dane używane dla innych aktywnych subskrypcji.",
   "Update payment details": "Update payment details",
   "Card number": "Card number",
   "Your payment details have been updated": "Your payment details have been updated",
@@ -305,7 +308,6 @@
   "Go back to the login page": "Go back to the login page",
   "Not sure that was the right email address?": "Not sure that was the right email address?",
   "Try again.": "Try again.",
-  "Thank You!": "Thank You!",
   "Your purchase has been successfully completed.": "Your purchase has been successfully completed.",
   "We hope you love it. If you need help from us with your account, you can always find it": "We hope you love it. If you need help from us with your account, you can always find it",
   "here": "here",
@@ -359,6 +361,7 @@
   "Here are the plans that might match your needs:": "Here are the plans that might match your needs:",
   "Or still wants to cancel a subscription?": "Or still wants to cancel a subscription?",
   "Cancel switch": "Cancel switch",
+  "Are you sure you want to cancel the switch?": "Are you sure you want to cancel the switch?",
   "Switch canceled": "Switch canceled",
   "Your {{switchDirection}} to {{switchOfferTitle}} is still pending and will take effect on {{baseOfferExpirationDate}}. If you decide to cancel the switch, you will keep access to current plan and be charged {{baseOfferPrice}} on the next billing date.": "Your {{switchDirection}} to {{switchOfferTitle}} is still pending and will take effect on {{baseOfferExpirationDate}}. If you decide to cancel the switch, you will keep access to current plan and be charged {{baseOfferPrice}} on the next billing date.",
   "You have successfully canceled your {{switchDirection}} to {{switchOfferTitle}}. You will be charged a current price on {{baseOfferExpirationDate}} and keep access to {{baseOfferTitle}}.": "You have successfully canceled your {{switchDirection}} to {{switchOfferTitle}}. You will be charged a current price on {{baseOfferExpirationDate}} and keep access to {{baseOfferTitle}}.",
